+++ sources/Google.Solutions.Common.Test/Util/TestDisposableContainer.cs
@@ -1,0 +1,71 @@
+﻿//
+// Copyright 2023 Google LLC
+//
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+// 
+//   http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+//
+
+using Google.Solutions.Common.Util;
+using Moq;
+using NUnit.Framework;
+using System;
+
+namespace Google.Solutions.Common.Test.Util
+{
+    [TestFixture]
+    public class TestDisposableContainer
+    {
+        [Test]
+        public void WhenEmpty_ThenDisposeSucceeds()
+        {
+            using (new DisposableContainer())
+            { }
+        }
+
+        //---------------------------------------------------------------------
+        // Add.
+        //---------------------------------------------------------------------
+
+        [Test]
+        public void AddNull()
+        {
+            using (var container = new DisposableContainer())
+            {
+                container.Add(null);
+            }
+        }
+
+        //---------------------------------------------------------------------
+        // Dispose.
+        //---------------------------------------------------------------------
+
+        [Test]
+        public void WhenNotEmpty_ThenDisposeDisposesAllObjects()
+        {
+            var disposable1 = new Mock<IDisposable>();
+            var disposable2 = new Mock<IDisposable>();
+            
+            using (var container = new DisposableContainer(disposable1.Object))
+            {
+                container.Add(disposable2.Object);
+            }
+
+            disposable1.Verify(d => d.Dispose(), Times.Once);
+            disposable2.Verify(d => d.Dispose(), Times.Once);
+        }
+    }
+}
